requirements.txt
fastapi==0.124.4
uvicorn==0.38.0

SQLAlchemy==2.0.36
greenlet>=3.0.0
asyncpg
aiosqlite
psycopg[binary]>=3.1


python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

python-dotenv==1.2.1
python-multipart==0.0.20

Jinja2==3.1.6
requests==2.32.5
pytz==2025.2
httpx==0.27.0
pydantic-settings